Move validate_csv() function to its own module
This is to avoid a circular dependency that was triggered when importing
the validate_csv method, resulting in the db module also being imported
(since it is used by the module where validate_csv() used to live).
